Tony fixes to matlab interface
Finished make test sys/os/glnxa64 thing (and it needed an export).
Updated the instructions in the site file now that it's been moved. I
played a little bit with cl and f2c - it almost worked, but not quite.
Turns out ssize_t is a GNU-ism but ptrdiff_t needs #include <cstddef>
for GCC 4.6 (otherwise "error: 'ptrdiff_t' does not name a type").
Shouldn't use libtool for cl, but when we do use libtool make clean
should delete the *.lo files. 



git-svn-id: https://projects.coin-or.org/svn/Ipopt/trunk@2158 ef2a96d0-92f8-0310-a3c1-cdaa70078a94